+++ ktor-openapi/src/main/kotlin/io/github/smiley4/ktoropenapi/config/RequestParameterConfig.kt
@@ -1,12 +1,14 @@
 package io.github.smiley4.ktoropenapi.config
 
 import io.github.smiley4.ktoropenapi.config.descriptors.ExampleDescriptor
+import io.github.smiley4.ktoropenapi.config.descriptors.KTypeDescriptor
 import io.github.smiley4.ktoropenapi.config.descriptors.SwaggerExampleDescriptor
 import io.github.smiley4.ktoropenapi.config.descriptors.ValueExampleDescriptor
 import io.github.smiley4.ktoropenapi.config.descriptors.TypeDescriptor
 import io.github.smiley4.ktoropenapi.data.RequestParameterData
 import io.swagger.v3.oas.models.examples.Example
 import io.swagger.v3.oas.models.parameters.Parameter
+import kotlin.reflect.typeOf
 
 /**
  * Describes a single request parameter.
@@ -56,15 +58,16 @@ class RequestParameterConfig internal constructor(
     /**
      * An example value for this parameter
      */
-    fun example(name: String, example: ValueExampleDescriptorConfig.() -> Unit) = example(
-        ValueExampleDescriptorConfig()
+    inline fun <reified T> example(name: String, example: ValueExampleDescriptorConfig<T>.() -> Unit) = example(
+        ValueExampleDescriptorConfig<T>()
             .apply(example)
             .let { result ->
                 ValueExampleDescriptor(
                     name = name,
                     value = result.value,
                     summary = result.summary,
-                    description = result.description
+                    description = result.description,
+                    type = KTypeDescriptor(typeOf<T>())
                 )
             }
     )