+++ ktor-openapi/src/main/kotlin/io/github/smiley4/ktoropenapi/config/SimpleBodyConfig.kt
@@ -1,12 +1,14 @@
 package io.github.smiley4.ktoropenapi.config
 
 import io.github.smiley4.ktoropenapi.config.descriptors.ExampleDescriptor
+import io.github.smiley4.ktoropenapi.config.descriptors.KTypeDescriptor
 import io.github.smiley4.ktoropenapi.config.descriptors.RefExampleDescriptor
 import io.github.smiley4.ktoropenapi.config.descriptors.SwaggerExampleDescriptor
 import io.github.smiley4.ktoropenapi.config.descriptors.ValueExampleDescriptor
 import io.github.smiley4.ktoropenapi.config.descriptors.TypeDescriptor
 import io.github.smiley4.ktoropenapi.data.SimpleBodyData
 import io.swagger.v3.oas.models.examples.Example
+import kotlin.reflect.typeOf
 
 
 /**
@@ -40,15 +42,16 @@ class SimpleBodyConfig internal constructor(
     /**
      * Add the given example as an example to this body
      */
-    fun example(name: String, example: ValueExampleDescriptorConfig.() -> Unit) = example(
-        ValueExampleDescriptorConfig()
+    inline fun <reified T> example(name: String, example: ValueExampleDescriptorConfig<T>.() -> Unit) = example(
+        ValueExampleDescriptorConfig<T>()
             .apply(example)
             .let { result ->
                 ValueExampleDescriptor(
                     name = name,
                     value = result.value,
                     summary = result.summary,
-                    description = result.description
+                    description = result.description,
+                    type = KTypeDescriptor(typeOf<T>())
                 )
             }
     )