+++ ktor-openapi/src/main/kotlin/io/github/smiley4/ktoropenapi/data/RouteData.kt
@@ -17,4 +17,5 @@ internal data class RouteData(
     val responses: List<ResponseData>,
     val externalDocs: ExternalDocsData?,
     val servers: List<ServerData>,
+    val extensions: Map<String, Any?>
 )